requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.9
django-import-export==2.6.1
django-js-asset==1.2.2
django-mdeditor==0.1.18
django-mptt==0.13.4
django-simpleui==2021.11.5
django-taggit==1.5.1
et-xmlfile==1.1.0
Markdown==3.3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
PyMySQL==1.0.2
python-memcached==1.59
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
xlrd==2.0.1
xlwt==1.3.0
